// legacy-convex-projects/fu/convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  audit: typeof audit;
  auth: typeof auth;
  authz: typeof authz;
  bookings: typeof bookings;
  crons: typeof crons;
  dashboard: typeof dashboard;
  errors: typeof errors;
  globalCategories: typeof globalCategories;
  http: typeof http;
  mediaManagement: typeof mediaManagement;
  myFunctions: typeof myFunctions;
  orders: typeof orders;
  organizationMembers: typeof organizationMembers;
  organizations: typeof organizations;
  providers: typeof providers;
  security: typeof security;
  sellerCategories: typeof sellerCategories;
  sellerCustomization: typeof sellerCustomization;
  sellerOrders: typeof sellerOrders;
  sellerProducts: typeof sellerProducts;
  services: typeof services;
  users: typeof users;
  utils: typeof utils;
  variants: typeof variants;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;